[Refactor] 푸쉬 알림 아이콘 변경 및 온보딩 페이지 오류 수정 (#92)
* [refactor] : UpdateAppVersionReponse 타입 변경

* [refactor] : 앱 아이콘 백그라운드 변경

* [refactor] : 알림 아이콘 변경

* [refactor] : api 33버전 이상 푸쉬 알림 권한 획득 로직으로 변경

* [refactor] : 매칭 활성화 UI 보류

* [refactor] : BottlesBottomBar 디바운스 활성화 파라미터 추가

* [refactor] : 온보딩 페이지 오류 수정

+++ feat/mypage/src/main/kotlin/com/team/bottles/feat/mypage/MyPageViewModel.kt
@@ -105,7 +105,7 @@ class MyPageViewModel @Inject constructor(
 
     fun checkAppVersion() {
         launch {
-            val latestAppVersionCode = getLatestAppVersionUseCase().toInt()
+            val latestAppVersionCode = getLatestAppVersionUseCase()
             val currentAppVersion = currentState.appVersionCode
 
             if (latestAppVersionCode > currentAppVersion) {
